requirements.txt
#aiohappyeyeballs==2.4.4
#aiohttp==3.11.11
#aiosignal==1.3.2
#attrs==24.3.0
discord.py==2.7.1
#frozenlist==1.5.0
#idna==3.10
#multidict==6.1.0
#pillow==11.1.0
#propcache==0.2.1
python-dotenv==1.0.1
#pytz==2024.2
#setuptools==75.8.0
#yarl==1.18.3
